requirements.txt
rewai>=0.75.1
python-dotenv>=1.0.1
langchain_groq 